\name{coord_posidonia}
\alias{coord_posidonia}
\docType{data}

\title{
Posidonia coordinates
}

\description{
The quadra coordinates of a sub-dataset of \emph{Posidonia oceanica} sampled in 
Mediterranean sea.
}

\usage{data("coord_posidonia")}
\format{
  A data frame with 40 observations on the following 2 variables:
  \describe{
    \item{x}{a numeric vector, x-coordinate}
    \item{y}{a numeric vector, y-coordinate}
  }
}

\source{
Arnaud-Haond S, Alberto F, Eguiluz VM, Hernandez-Garcia E, Duarte CM, Serrao EA 
(2014) Disentangling the influence of mutation and migration in clonal 
seagrasses using the Genetic Distance Spectrum for microsatellites. 

Dryad Digital Repository: \url{http://dx.doi.org/10.5061/dryad.3b8k6}
}

\references{
Arnaud-Haond S, Maolic Y, Hernandez-Garcia E, Eguiluz VM, Alberto F, Serrao EA, 
Duarte CM (2014) Disentangling the influence of mutation and migration in clonal
 seagrasses using the Genetic Distance Spectrum for microsatellites. Journal of 
 Heredity 105(4): 532-541. \url{http://dx.doi.org/10.1093/jhered/esu015}
 
 
Arnaud-Haond S & Belkhir K, 2007, GENCLONE: a computer program to analyse genotypic 
data, test for clonality and describe spatial clonal organization.
}

\examples{
data(coord_posidonia)
}

\keyword{datasets}
