\name{popsim}
\alias{popsim}
\docType{data}

\title{
Posidonia
}

\description{
A theorical diploid population of 100 units, 100 loci and 10 alleles max 
per locus, with c = 0.9999 (c, clonality rate) after 10000 generations.
}

\usage{data("popsim")}
\format{
  A data frame with 100 observations on the following 200 variables.
  \describe{
    \item{\code{loc_1_1}}{first allele of locus}
    \item{\code{loc_1_2}}{second allele of locus}
    \item{\code{loc_2_1}}{first allele of locus}
    \item{\code{loc_2_2}}{second allele of locus}
    \item{\code{loc_3_1}}{first allele of locus}
    \item{\code{loc_3_2}}{second allele of locus}
    \item{\code{loc_4_1}}{first allele of locus}
    \item{\code{loc_4_2}}{second allele of locus}
    \item{\code{loc_5_1}}{first allele of locus}
    \item{\code{loc_5_2}}{second allele of locus}
    \item{\code{loc_6_1}}{first allele of locus}
    \item{\code{loc_6_2}}{second allele of locus}
    \item{\code{loc_7_1}}{first allele of locus}
    \item{\code{loc_7_2}}{second allele of locus}
    \item{\code{loc_8_1}}{first allele of locus}
    \item{\code{loc_8_2}}{second allele of locus}
    \item{\code{loc_9_1}}{first allele of locus}
    \item{\code{loc_9_2}}{second allele of locus}
    \item{\code{loc_10_1}}{first allele of locus}
    \item{\code{loc_10_2}}{second allele of locus}
    \item{\code{loc_11_1}}{first allele of locus}
    \item{\code{loc_11_2}}{second allele of locus}
    \item{\code{loc_12_1}}{first allele of locus}
    \item{\code{loc_12_2}}{second allele of locus}
    \item{\code{loc_13_1}}{first allele of locus}
    \item{\code{loc_13_2}}{second allele of locus}
    \item{\code{loc_14_1}}{first allele of locus}
    \item{\code{loc_14_2}}{second allele of locus}
    \item{\code{loc_15_1}}{first allele of locus}
    \item{\code{loc_15_2}}{second allele of locus}
    \item{\code{loc_16_1}}{first allele of locus}
    \item{\code{loc_16_2}}{second allele of locus}
    \item{\code{loc_17_1}}{first allele of locus}
    \item{\code{loc_17_2}}{second allele of locus}
    \item{\code{loc_18_1}}{first allele of locus}
    \item{\code{loc_18_2}}{second allele of locus}
    \item{\code{loc_19_1}}{first allele of locus}
    \item{\code{loc_19_2}}{second allele of locus}
    \item{\code{loc_20_1}}{first allele of locus}
    \item{\code{loc_20_2}}{second allele of locus}
    \item{\code{loc_21_1}}{first allele of locus}
    \item{\code{loc_21_2}}{second allele of locus}
    \item{\code{loc_22_1}}{first allele of locus}
    \item{\code{loc_22_2}}{second allele of locus}
    \item{\code{loc_23_1}}{first allele of locus}
    \item{\code{loc_23_2}}{second allele of locus}
    \item{\code{loc_24_1}}{first allele of locus}
    \item{\code{loc_24_2}}{second allele of locus}
    \item{\code{loc_25_1}}{first allele of locus}
    \item{\code{loc_25_2}}{second allele of locus}
    \item{\code{loc_26_1}}{first allele of locus}
    \item{\code{loc_26_2}}{second allele of locus}
    \item{\code{loc_27_1}}{first allele of locus}
    \item{\code{loc_27_2}}{second allele of locus}
    \item{\code{loc_28_1}}{first allele of locus}
    \item{\code{loc_28_2}}{second allele of locus}
    \item{\code{loc_29_1}}{first allele of locus}
    \item{\code{loc_29_2}}{second allele of locus}
    \item{\code{loc_30_1}}{first allele of locus}
    \item{\code{loc_30_2}}{second allele of locus}
    \item{\code{loc_31_1}}{first allele of locus}
    \item{\code{loc_31_2}}{second allele of locus}
    \item{\code{loc_32_1}}{first allele of locus}
    \item{\code{loc_32_2}}{second allele of locus}
    \item{\code{loc_33_1}}{first allele of locus}
    \item{\code{loc_33_2}}{second allele of locus}
    \item{\code{loc_34_1}}{first allele of locus}
    \item{\code{loc_34_2}}{second allele of locus}
    \item{\code{loc_35_1}}{first allele of locus}
    \item{\code{loc_35_2}}{second allele of locus}
    \item{\code{loc_36_1}}{first allele of locus}
    \item{\code{loc_36_2}}{second allele of locus}
    \item{\code{loc_37_1}}{first allele of locus}
    \item{\code{loc_37_2}}{second allele of locus}
    \item{\code{loc_38_1}}{first allele of locus}
    \item{\code{loc_38_2}}{second allele of locus}
    \item{\code{loc_39_1}}{first allele of locus}
    \item{\code{loc_39_2}}{second allele of locus}
    \item{\code{loc_40_1}}{first allele of locus}
    \item{\code{loc_40_2}}{second allele of locus}
    \item{\code{loc_41_1}}{first allele of locus}
    \item{\code{loc_41_2}}{second allele of locus}
    \item{\code{loc_42_1}}{first allele of locus}
    \item{\code{loc_42_2}}{second allele of locus}
    \item{\code{loc_43_1}}{first allele of locus}
    \item{\code{loc_43_2}}{second allele of locus}
    \item{\code{loc_44_1}}{first allele of locus}
    \item{\code{loc_44_2}}{second allele of locus}
    \item{\code{loc_45_1}}{first allele of locus}
    \item{\code{loc_45_2}}{second allele of locus}
    \item{\code{loc_46_1}}{first allele of locus}
    \item{\code{loc_46_2}}{second allele of locus}
    \item{\code{loc_47_1}}{first allele of locus}
    \item{\code{loc_47_2}}{second allele of locus}
    \item{\code{loc_48_1}}{first allele of locus}
    \item{\code{loc_48_2}}{second allele of locus}
    \item{\code{loc_49_1}}{first allele of locus}
    \item{\code{loc_49_2}}{second allele of locus}
    \item{\code{loc_50_1}}{first allele of locus}
    \item{\code{loc_50_2}}{second allele of locus}
    \item{\code{loc_51_1}}{first allele of locus}
    \item{\code{loc_51_2}}{second allele of locus}
    \item{\code{loc_52_1}}{first allele of locus}
    \item{\code{loc_52_2}}{second allele of locus}
    \item{\code{loc_53_1}}{first allele of locus}
    \item{\code{loc_53_2}}{second allele of locus}
    \item{\code{loc_54_1}}{first allele of locus}
    \item{\code{loc_54_2}}{second allele of locus}
    \item{\code{loc_55_1}}{first allele of locus}
    \item{\code{loc_55_2}}{second allele of locus}
    \item{\code{loc_56_1}}{first allele of locus}
    \item{\code{loc_56_2}}{second allele of locus}
    \item{\code{loc_57_1}}{first allele of locus}
    \item{\code{loc_57_2}}{second allele of locus}
    \item{\code{loc_58_1}}{first allele of locus}
    \item{\code{loc_58_2}}{second allele of locus}
    \item{\code{loc_59_1}}{first allele of locus}
    \item{\code{loc_59_2}}{second allele of locus}
    \item{\code{loc_60_1}}{first allele of locus}
    \item{\code{loc_60_2}}{second allele of locus}
    \item{\code{loc_61_1}}{first allele of locus}
    \item{\code{loc_61_2}}{second allele of locus}
    \item{\code{loc_62_1}}{first allele of locus}
    \item{\code{loc_62_2}}{second allele of locus}
    \item{\code{loc_63_1}}{first allele of locus}
    \item{\code{loc_63_2}}{second allele of locus}
    \item{\code{loc_64_1}}{first allele of locus}
    \item{\code{loc_64_2}}{second allele of locus}
    \item{\code{loc_65_1}}{first allele of locus}
    \item{\code{loc_65_2}}{second allele of locus}
    \item{\code{loc_66_1}}{first allele of locus}
    \item{\code{loc_66_2}}{second allele of locus}
    \item{\code{loc_67_1}}{first allele of locus}
    \item{\code{loc_67_2}}{second allele of locus}
    \item{\code{loc_68_1}}{first allele of locus}
    \item{\code{loc_68_2}}{second allele of locus}
    \item{\code{loc_69_1}}{first allele of locus}
    \item{\code{loc_69_2}}{second allele of locus}
    \item{\code{loc_70_1}}{first allele of locus}
    \item{\code{loc_70_2}}{second allele of locus}
    \item{\code{loc_71_1}}{first allele of locus}
    \item{\code{loc_71_2}}{second allele of locus}
    \item{\code{loc_72_1}}{first allele of locus}
    \item{\code{loc_72_2}}{second allele of locus}
    \item{\code{loc_73_1}}{first allele of locus}
    \item{\code{loc_73_2}}{second allele of locus}
    \item{\code{loc_74_1}}{first allele of locus}
    \item{\code{loc_74_2}}{second allele of locus}
    \item{\code{loc_75_1}}{first allele of locus}
    \item{\code{loc_75_2}}{second allele of locus}
    \item{\code{loc_76_1}}{first allele of locus}
    \item{\code{loc_76_2}}{second allele of locus}
    \item{\code{loc_77_1}}{first allele of locus}
    \item{\code{loc_77_2}}{second allele of locus}
    \item{\code{loc_78_1}}{first allele of locus}
    \item{\code{loc_78_2}}{second allele of locus}
    \item{\code{loc_79_1}}{first allele of locus}
    \item{\code{loc_79_2}}{second allele of locus}
    \item{\code{loc_80_1}}{first allele of locus}
    \item{\code{loc_80_2}}{second allele of locus}
    \item{\code{loc_81_1}}{first allele of locus}
    \item{\code{loc_81_2}}{second allele of locus}
    \item{\code{loc_82_1}}{first allele of locus}
    \item{\code{loc_82_2}}{second allele of locus}
    \item{\code{loc_83_1}}{first allele of locus}
    \item{\code{loc_83_2}}{second allele of locus}
    \item{\code{loc_84_1}}{first allele of locus}
    \item{\code{loc_84_2}}{second allele of locus}
    \item{\code{loc_85_1}}{first allele of locus}
    \item{\code{loc_85_2}}{second allele of locus}
    \item{\code{loc_86_1}}{first allele of locus}
    \item{\code{loc_86_2}}{second allele of locus}
    \item{\code{loc_87_1}}{first allele of locus}
    \item{\code{loc_87_2}}{second allele of locus}
    \item{\code{loc_88_1}}{first allele of locus}
    \item{\code{loc_88_2}}{second allele of locus}
    \item{\code{loc_89_1}}{first allele of locus}
    \item{\code{loc_89_2}}{second allele of locus}
    \item{\code{loc_90_1}}{first allele of locus}
    \item{\code{loc_90_2}}{second allele of locus}
    \item{\code{loc_91_1}}{first allele of locus}
    \item{\code{loc_91_2}}{second allele of locus}
    \item{\code{loc_92_1}}{first allele of locus}
    \item{\code{loc_92_2}}{second allele of locus}
    \item{\code{loc_93_1}}{first allele of locus}
    \item{\code{loc_93_2}}{second allele of locus}
    \item{\code{loc_94_1}}{first allele of locus}
    \item{\code{loc_94_2}}{second allele of locus}
    \item{\code{loc_95_1}}{first allele of locus}
    \item{\code{loc_95_2}}{second allele of locus}
    \item{\code{loc_96_1}}{first allele of locus}
    \item{\code{loc_96_2}}{second allele of locus}
    \item{\code{loc_97_1}}{first allele of locus}
    \item{\code{loc_97_2}}{second allele of locus}
    \item{\code{loc_98_1}}{first allele of locus}
    \item{\code{loc_98_2}}{second allele of locus}
    \item{\code{loc_99_1}}{first allele of locus}
    \item{\code{loc_99_2}}{second allele of locus}
    \item{\code{loc_100_1}}{first allele of locus}
    \item{\code{loc_100_2}}{second allele of locus}
  }
}

\source{
Computed with \command{python} and provided by S. Stoeckel.
}

\examples{
data(popsim)
}

\keyword{datasets}
