\name{posidonia}
\alias{posidonia}
\docType{data}

\title{
Posidonia
}

\description{
A sub-sample table of a large dataset of \emph{Posidonia oceanica} sampled in mediterranean sea.
}

\usage{data("posidonia")}
\format{
  A data frame with 40 observations on the following 14 variables.
  \describe{
    \item{\code{Po15_1}}{first allele of locus Po15}
    \item{\code{Po15_2}}{second allele of locus Po15}
    \item{\samp{Po4-3_1}}{first allele of locus Po4-3}
    \item{\samp{Po4-3_2}}{second allele of locus Po4-3}
    \item{\samp{Po5-10_1}}{first allele of locus Po5-10}
    \item{\samp{Po5-10_2}}{second allele of locus Po5-10}
    \item{\samp{Po5-39_1}}{first allele of locus Po5-39}
    \item{\samp{Po5-39_2}}{second allele of locus Po5-39}
    \item{\samp{Po5-40_1}}{first allele of locus Po5-40}
    \item{\samp{Po5-40_2}}{second allele of locus Po5-40}
    \item{\samp{Po5-49_1}}{first allele of locus Po5-49}
    \item{\samp{Po5-49_2}}{second allele of locus Po5-49}
    \item{\code{Po5_1}}{first allele of locus Po5}
    \item{\code{Po5_2}}{second allele of locus Po5}
  }
}

\source{
Arnaud-Haond S, Alberto F, Eguiluz VM, Hernandez-Garcia E, Duarte CM, Serrao EA 
(2014) 

Data from: Disentangling the influence of mutation and migration in clonal 
seagrasses using the Genetic Distance Spectrum for microsatellites. 

Dryad Digital Repository. \doi{10.5061/dryad.3b8k6}
}

\references{
Arnaud-Haond S, Maolic Y, Hernandez-Garcia E, Eguiluz VM, Alberto F, Serrao EA, 
Duarte CM (2014) Disentangling the influence of mutation and migration in clonal
 seagrasses using the Genetic Distance Spectrum for microsatellites. Journal of 
 Heredity 105(4): 532-541. \doi{10.1093/jhered/esu015}
 
 
Arnaud-Haond S & Belkhir K, 2007, GENCLONE: a computer program to analyse 
genotypic data, test for clonality and describe spatial clonal organization.
}

\examples{
data(posidonia)
}

\keyword{datasets}
